% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdist.R
\name{pdist}
\alias{pdist}
\title{Distance Measures between Multiple Empirical Cumulative Distribution Functions}
\usage{
pdist(elist, type = c("KS", "Lp", "Wasserstein"), p = 2, as.dist = FALSE)
}
\arguments{
\item{elist}{a length \eqn{N} list of \code{ecdf} objects.}

\item{type}{name of the distance/dissimilarity measure. Case insensitive.}

\item{p}{exponent for \code{Lp} or \code{Wasserstein} distance.}

\item{as.dist}{a logical; \code{TRUE} to return \code{dist} object, \code{FALSE} to return an \eqn{(N\times N)} symmetric matrix of pairwise distances.}
}
\value{
either \code{dist} object of an \eqn{(N\times N)} symmetric matrix of pairwise distances by \code{as.dist} argument.
}
\description{
We measure distance between two empirical cumulative distribution functions (ECDF). For 
simplicity, we only take an input of \code{\link[stats]{ecdf}} objects from \pkg{stats} package.
}
\examples{
## toy example : 10 of random and uniform distributions
mylist = list()
for (i in 1:10){
  mylist[[i]] = stats::ecdf(stats::rnorm(50, sd=2))
}
for (i in 11:20){
  mylist[[i]] = stats::ecdf(stats::runif(50, min=-5))
}

## compute several distance measures
dk = pdist(mylist, type="KS")          # Kolmogorov-Smirnov
dl = pdist(mylist, type="Lp")          # L2 
dw = pdist(mylist, type="Wasserstein") # 2-Wasserstein

## visualize
opar = par(mfrow=c(1,3), pty="s")
image(dk[,nrow(dk):1], axes=FALSE, main="Kolmogorov-Smirnov")
image(dl[,nrow(dl):1], axes=FALSE, main="L2")
image(dw[,nrow(dw):1], axes=FALSE, main="2-Wasserstein")
par(opar)

}
\seealso{
\code{\link[stats]{ecdf}}
}
